requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-bootstrap4==3.0.1
django-crispy-forms==1.11.2
django-embed-video==1.4.0
idna==2.10
mysqlclient==2.0.3
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.5
